@@ -1,7 +1,7 @@
 import React from "react";
 import { Provider } from "react-redux";
 import { configureStore } from "../store";
-import { BrowserRouter as Router } from "react-router-dom";
+import { BrowserRouter } from "react-router-dom";
 import Navbar from "./Navbar";
 import Main from "./Main";
 
@@ -14,12 +14,10 @@ const store = configureStore();
 
 const App = () => (
   <Provider store={store}>
-    <Router>
-      <div className="onboarding">
-        <Navbar />
-        <Main />
-      </div>
-    </Router>
+    <BrowserRouter>
+      <Navbar />
+      <Main />
+    </BrowserRouter>
   </Provider>
 );
 
